+++ spx_libs/l_nvm.h
@@ -14,8 +14,8 @@
 #include "stdlib.h"
 #include "stdbool.h"
 
-int8_t NVMEE_read( uint32_t rdAddress, char *data, uint8_t length );
-int8_t NVMEE_write( uint32_t wrAddress, char *data, uint8_t length );
+int8_t NVMEE_read( uint32_t rdAddress, char *data, uint16_t length );
+int8_t NVMEE_write( uint32_t wrAddress, char *data, uint16_t length );
 char *NVMEE_readID(void);
 
 #endif /* SPX_LIBS_L_NVM_H_ */
